+++ app/Jobs/CheckDomainJob.php
@@ -13,6 +13,10 @@ class CheckDomainJob implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
+    public int $tries = 2;
+
+    public int $timeout = 120;
+
     public function __construct(
         public Domain $domain,
         public bool $synchronous = false,