+++ app/Jobs/SendMonthlyReportsJob.php
@@ -17,6 +17,10 @@ class SendMonthlyReportsJob implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
+    public int $tries = 2;
+
+    public int $timeout = 600;
+
     /**
      * Execute the job.
      */